+++ packages/shared/src/index.server.ts
@@ -20,15 +20,16 @@ export {
 } from "./types.js";
 export {
     base64Decode,
-    loadConfig,
     loadJsonFile,
-    isRemotePath,
     getConfigSettings,
+    getRepoPath,
 } from "./utils.js";
 export * from "./constants.js";
 export {
     env,
     resolveEnvironmentVariableOverridesFromConfig,
+    loadConfig,
+    isRemotePath,
 } from "./env.server.js";
 export {
     createLogger,